TST: stats: move tests into classes
